添加查看和创建节点信息功能

+++ zkGui/MainForm.cs
@@ -36,6 +36,11 @@ namespace zkGui
             return $"{parentPath.TrimEnd('/')}/{childPath.TrimStart('/')}";
         }
 
+        private string GetNodeName(string path)
+        {
+            return path.Split('/').LastOrDefault();
+        }
+
         private void SelecteNode(string name)
         {
             treeViewNodes.SelectedNode = treeViewNodes.Nodes.Find(name, true).FirstOrDefault();
@@ -79,7 +84,14 @@ namespace zkGui
             if (m_closing)
             {
                 m_stateWatcher.OnEvent -= M_stateWatcher_OnEvent;
-                await m_zooKeeperClient.closeAsync();
+                try
+                {
+                    await m_zooKeeperClient.closeAsync();
+                }
+                catch (KeeperException ex)
+                {
+                    Log(ex.Message);
+                }
                 m_zooKeeperClient = null;
                 buttonConnect.BackColor = SystemColors.Control;
                 buttonConnect.Text = "Connect";
@@ -87,7 +99,7 @@ namespace zkGui
             }
         }
 
-        private async void M_stateWatcher_OnEvent(WatchedEvent obj)
+        private void M_stateWatcher_OnEvent(WatchedEvent obj)
         {
             Color color = Color.Gray;
             Log(obj.getState().ToString());
@@ -174,7 +186,14 @@ namespace zkGui
                     var dialogResult = authInfoForm.ShowDialog();
                     if (dialogResult == DialogResult.OK)
                     {
-                        m_zooKeeperClient.addAuthInfo(authInfoForm.Scheme, authInfoForm.Auth);
+                        try
+                        {
+                            m_zooKeeperClient.addAuthInfo(authInfoForm.Scheme, authInfoForm.Auth);
+                        }
+                        catch (KeeperException ex)
+                        {
+                            Log(ex.Message);
+                        }
                     }
                 }
             }
@@ -184,7 +203,50 @@ namespace zkGui
         {
             if (m_zooKeeperClient != null)
             {
-                await m_zooKeeperClient?.closeAsync();
+                try
+                {
+                    await m_zooKeeperClient.closeAsync();
+                }
+                catch (KeeperException ex)
+                {
+                    Log(ex.Message);
+                }
+            }
+        }
+
+        private async void buttonCreate_Click(object sender, EventArgs e)
+        {
+            var selectedNode = treeViewNodes.SelectedNode;
+            if (selectedNode != null)
+            {
+                NodeInfoForm nodeInfoForm = new NodeInfoForm();
+                nodeInfoForm.NodeName = "NewNode";
+                nodeInfoForm.Text = "NewNode";
+                nodeInfoForm.ACLs = new List<org.apache.zookeeper.data.ACL>() { new org.apache.zookeeper.data.ACL(31, new org.apache.zookeeper.data.Id("world", "anyone")) };
+                nodeInfoForm.ShowDialog();
+                if (MessageBox.Show($"确认添加节点{nodeInfoForm.NodeName}？", "确认", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                    try
+                    {
+                        var data = nodeInfoForm.Data;
+                        var acls = nodeInfoForm.ACLs;
+                        string path = await m_zooKeeperClient.createAsync(Combine(selectedNode.Name, nodeInfoForm.NodeName), data, acls, nodeInfoForm.CreateMode);
+                        selectedNode.Nodes.Add(path, GetNodeName(path));
+                        Log($"添加成功 {path}");
+                    }
+                    catch (KeeperException ex)
+                    {
+                        Log($"{ex.Message} {Combine(selectedNode.Name, nodeInfoForm.NodeName)}");
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message, "添加节点异常");
+                    }
+                }
+            }
+            else
+            {
+                MessageBox.Show("未选择节点", "提示");
             }
         }
 
@@ -193,11 +255,79 @@ namespace zkGui
             var selectedNode = treeViewNodes.SelectedNode;
             if (selectedNode != null)
             {
+                if (MessageBox.Show($"确认删除节点{selectedNode.Text}？", "确认", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                    try
+                    {
+                        await m_zooKeeperClient.deleteAsync(selectedNode.Name);
+                        SelecteNode(selectedNode.Parent.Name);
+                        selectedNode.Parent?.Nodes.Remove(selectedNode);
+                        Log($"删除成功 {selectedNode.Name}");
+                    }
+                    catch (KeeperException ex)
+                    {
+                        Log($"{ex.Message} {selectedNode.Name}");
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message, "删除节点异常");
+                    }
+                }
+            }
+            else
+            {
+                MessageBox.Show("未选择节点", "提示");
+            }
+        }
+
+        private async void buttonInfo_Click(object sender, EventArgs e)
+        {
+            var selectedNode = treeViewNodes.SelectedNode;
+            if (selectedNode != null)
+            {
                 try
                 {
-                    await m_zooKeeperClient.deleteAsync(selectedNode.Name);
-                    SelecteNode(selectedNode.Parent.Name);
-                    selectedNode.Parent?.Nodes.Remove(selectedNode);
+                    var dataResult = await m_zooKeeperClient.getDataAsync(selectedNode.Name);
+                    var aCLResult = await m_zooKeeperClient.getACLAsync(selectedNode.Name);
+                    NodeInfoForm nodeInfoForm = new NodeInfoForm();
+                    nodeInfoForm.NodeName = selectedNode.Text;
+                    nodeInfoForm.Text = selectedNode.Name;
+                    nodeInfoForm.Data = dataResult.Data;
+                    nodeInfoForm.ACLs = aCLResult.Acls;
+                    nodeInfoForm.Stat = aCLResult.Stat;
+                    nodeInfoForm.ShowDialog();
+
+                    bool dataChanged = true;
+                    bool aclsChanged = true;
+                    var data = nodeInfoForm.Data;
+                    if (Convert.ToBase64String(dataResult.Data) == Convert.ToBase64String(data))
+                    {
+                        dataChanged = false;
+                    }
+
+                    var acls = nodeInfoForm.ACLs;
+                    if (aCLResult.Acls.Count == acls.Count && aCLResult.Acls.All(acl => acls.Exists(i => i.getPerms() == acl.getPerms() && i.getId().getScheme() == acl.getId().getScheme() && i.getId().getId() == acl.getId().getId())))
+                    {
+                        aclsChanged = false;
+                    }
+
+                    if (dataChanged || aclsChanged)
+                    {
+                        if (MessageBox.Show("是否要更新修改到zookeeper服务", "确认", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                        {
+                            if (dataChanged)
+                            {
+                                await m_zooKeeperClient.setDataAsync(selectedNode.Name, data);
+                                Log($"Data更新成功 {selectedNode.Name}");
+                            }
+
+                            if (aclsChanged)
+                            {
+                                await m_zooKeeperClient.setACLAsync(selectedNode.Name, acls);
+                                Log($"Acls更新成功 {selectedNode.Name}");
+                            }
+                        }
+                    }
                 }
                 catch (KeeperException ex)
                 {
@@ -205,7 +335,7 @@ namespace zkGui
                 }
                 catch (Exception ex)
                 {
-                    MessageBox.Show(ex.Message, "删除节点异常");
+                    MessageBox.Show(ex.Message, "显示节点信息异常");
                 }
             }
             else
